--- NeoLang/src/main/java/io/neolang/ast/NeoLangEOFToken.kt
@@ -1,6 +1,0 @@
-package io.neolang.ast
-
-/**
- * @author kiva
- */
-class NeoLangEOFToken : NeoLangToken(NeoLangTokenType.EOF, NeoLangTokenValue.EOF)
