--- NeoLang/src/main/java/io/neolang/ast/NeoLangToken.kt
@@ -1,13 +1,0 @@
-package io.neolang.ast
-
-/**
- * @author kiva
- */
-
-open class NeoLangToken(val tokenType: NeoLangTokenType, val tokenValue: NeoLangTokenValue) {
-  var lineNumber = 0
-
-  override fun toString(): String {
-    return "Token { tokenType: $tokenType, tokenValue: $tokenValue };"
-  }
-}
