--- NeoLang/src/main/java/io/neolang/ast/base/NeoLangAst.kt
@@ -1,12 +1,0 @@
-package io.neolang.ast.base
-
-import io.neolang.ast.visitor.VisitorFactory
-
-/**
- * @author kiva
- */
-open class NeoLangAst {
-  fun visit(): VisitorFactory {
-    return VisitorFactory(this)
-  }
-}
