--- NeoLang/src/main/java/io/neolang/ast/base/NeoLangBaseNode.kt
@@ -1,3 +1,0 @@
-package io.neolang.ast.base
-
-open class NeoLangBaseNode : NeoLangAst()
