--- NeoLang/src/main/java/io/neolang/ast/node/NeoLangBlockNode.kt
@@ -1,14 +1,0 @@
-package io.neolang.ast.node
-
-import io.neolang.ast.base.NeoLangBaseNode
-
-/**
- * @author kiva
- */
-class NeoLangBlockNode(blockElement: NeoLangBaseNode) : NeoLangAstBasedNode(blockElement) {
-  companion object {
-    fun emptyNode(): NeoLangBlockNode {
-      return NeoLangBlockNode(NeoLangDummyNode())
-    }
-  }
-}
