--- NeoLang/src/main/java/io/neolang/ast/node/NeoLangDummyNode.kt
@@ -1,8 +1,0 @@
-package io.neolang.ast.node
-
-import io.neolang.ast.base.NeoLangBaseNode
-
-/**
- * @author kiva
- */
-class NeoLangDummyNode : NeoLangBaseNode()
