--- NeoLang/src/main/java/io/neolang/ast/node/NeoLangNumberNode.kt
@@ -1,8 +1,0 @@
-package io.neolang.ast.node
-
-import io.neolang.ast.NeoLangToken
-
-/**
- * @author kiva
- */
-class NeoLangNumberNode(token: NeoLangToken) : NeoLangTokenBasedNode(token)
