--- NeoLang/src/main/java/io/neolang/ast/node/NeoLangStringNode.kt
@@ -1,8 +1,0 @@
-package io.neolang.ast.node
-
-import io.neolang.ast.NeoLangToken
-
-/**
- * @author kiva
- */
-class NeoLangStringNode(token: NeoLangToken) : NeoLangTokenBasedNode(token)
