--- NeoLang/src/main/java/io/neolang/ast/visitor/IVisitorCallback.kt
@@ -1,18 +1,0 @@
-package io.neolang.ast.visitor
-
-import io.neolang.runtime.context.NeoLangContext
-
-/**
- * @author kiva
- */
-interface IVisitorCallback {
-  fun onStart()
-
-  fun onFinish()
-
-  fun onEnterContext(contextName: String)
-
-  fun onExitContext()
-
-  fun getCurrentContext(): NeoLangContext
-}
