+++ NeoLang/src/main/java/io/neolang/frontend/visitors.kt
@@ -1,9 +1,9 @@
-package io.neolang.visitor
+package io.neolang.frontend
 
-import io.neolang.ast.visitor.IVisitorCallback
-import io.neolang.runtime.context.NeoLangContext
-import io.neolang.runtime.type.NeoLangArray
-import io.neolang.runtime.type.NeoLangValue
+import io.neolang.runtime.NeoLangArray
+import io.neolang.runtime.NeoLangContext
+import io.neolang.runtime.NeoLangValue
+import java.util.*
 
 class ConfigVisitor : IVisitorCallback {
   private var rootContext: NeoLangContext? = null
@@ -69,5 +69,42 @@ class ConfigVisitor : IVisitorCallback {
 
   override fun getCurrentContext(): NeoLangContext {
     return currentContext!!
+  }
+}
+
+/**
+ * @author kiva
+ */
+class DisplayProcessVisitor : IVisitorCallbackAdapter() {
+  private val contextStack = Stack<NeoLangContext>()
+
+  override fun onStart() {
+    println(">>> Start")
+    onEnterContext("global")
+  }
+
+  override fun onFinish() {
+    while (contextStack.isNotEmpty()) {
+      onExitContext()
+    }
+    println(">>> Finish")
+  }
+
+  override fun onEnterContext(contextName: String) {
+    val context = NeoLangContext(contextName)
+    contextStack.push(context)
+    println(">>> Entering Context `$contextName'")
+  }
+
+  override fun onExitContext() {
+    val context = contextStack.pop()
+    println(">>> Exiting & Dumping Context ${context.contextName}")
+    context.getAttributes().entries.forEach {
+      println("     > [${it.key}]: ${it.value.asString()}")
+    }
+  }
+
+  override fun getCurrentContext(): NeoLangContext {
+    return contextStack.peek()
   }
 }
