--- NeoLang/src/main/java/io/neolang/parser/InvalidTokenException.kt
@@ -1,6 +1,0 @@
-package io.neolang.parser
-
-/**
- * @author kiva
- */
-open class InvalidTokenException(message: String) : ParseException(message)
