--- NeoLang/src/main/java/io/neolang/parser/ParseException.kt
@@ -1,6 +1,0 @@
-package io.neolang.parser
-
-/**
- * @author kiva
- */
-open class ParseException(message: String) : RuntimeException(message)
