+++ NeoLang/src/main/java/io/neolang/runtime/types.kt
@@ -1,6 +1,4 @@
-package io.neolang.runtime.type
-
-import io.neolang.runtime.context.NeoLangContext
+package io.neolang.runtime
 
 /**
  * @author kiva
@@ -58,5 +56,51 @@ class NeoLangArray private constructor(
 
   override fun iterator(): Iterator<NeoLangArrayElement> {
     return elements.iterator()
+  }
+}
+
+/**
+ * @author kiva
+ */
+open class NeoLangArrayElement {
+  open fun eval(): NeoLangValue {
+    return NeoLangValue.UNDEFINED
+  }
+
+  open fun eval(key: String): NeoLangValue {
+    return NeoLangValue.UNDEFINED
+  }
+
+  open fun isBlock(): Boolean {
+    return false
+  }
+}
+
+/**
+ * @author kiva
+ */
+class NeoLangValue(private val rawValue: Any) {
+  fun asString(): String {
+    return rawValue.toString()
+  }
+
+  fun asNumber(): Double {
+    if (rawValue is Array<*>) {
+      return 0.0
+    }
+
+    try {
+      return rawValue.toString().toDouble()
+    } catch (e: Throwable) {
+      return 0.0
+    }
+  }
+
+  fun isValid(): Boolean {
+    return this != UNDEFINED
+  }
+
+  companion object {
+    val UNDEFINED = NeoLangValue("<undefined>")
   }
 }
