+++ app/src/main/java/io/neoterm/component/comp.kt
@@ -1,5 +1,18 @@
 package io.neoterm.component
 
+import android.content.Context
+import io.neoterm.component.codegen.CodeGenComponent
+import io.neoterm.component.colorscheme.ColorSchemeComponent
+import io.neoterm.component.completion.CompletionComponent
+import io.neoterm.component.config.ConfigureComponent
+import io.neoterm.component.extrakey.ExtraKeyComponent
+import io.neoterm.component.font.FontComponent
+import io.neoterm.component.pm.PackageComponent
+import io.neoterm.component.profile.ProfileComponent
+import io.neoterm.component.session.SessionComponent
+import io.neoterm.component.session.ShellProfile
+import io.neoterm.component.userscript.UserScriptComponent
+import io.neoterm.utils.NLog
 import java.util.concurrent.ConcurrentHashMap
 
 interface NeoComponent {
@@ -50,3 +63,28 @@ object ComponentManager {
 class ComponentDuplicateException(serviceName: String) : RuntimeException("Service $serviceName duplicate")
 class ComponentNotFoundException(serviceName: String) : RuntimeException("Component `$serviceName' not found")
 
+/**
+ * @author kiva
+ */
+object NeoInitializer {
+  fun init(context: Context) {
+    NLog.init(context)
+    initComponents()
+  }
+
+  fun initComponents() {
+    ComponentManager.registerComponent(ConfigureComponent::class.java)
+    ComponentManager.registerComponent(CodeGenComponent::class.java)
+    ComponentManager.registerComponent(ColorSchemeComponent::class.java)
+    ComponentManager.registerComponent(FontComponent::class.java)
+    ComponentManager.registerComponent(UserScriptComponent::class.java)
+    ComponentManager.registerComponent(ExtraKeyComponent::class.java)
+    ComponentManager.registerComponent(CompletionComponent::class.java)
+    ComponentManager.registerComponent(PackageComponent::class.java)
+    ComponentManager.registerComponent(SessionComponent::class.java)
+    ComponentManager.registerComponent(ProfileComponent::class.java)
+
+    val profileComp = ComponentManager.getComponent<ProfileComponent>()
+    profileComp.registerProfile(ShellProfile.PROFILE_META_NAME, ShellProfile::class.java)
+  }
+}
